InputManager & DisplayManager Created
Created a class to handle user input using a scanner. This class also has a method that can be called in catch blocks to handle the exception and output a message to the user. DisplayManager contains a class for outputting a welcome message and stats to the user. Further implementation and testing will be carried out on both classes.