Class To Convert Textual Numbers To Numbers
As part of the projects UI I have implemented a class that can take a string based textual number inputs (fifty five months) and convert them into numbers for processing (55 months). This class consists of methods that split the string, cleans the string of punctuation and unwanted words, replaces the textual numbers with numbers and then pieces the string back together with the loose words and punctuation. This will be expanded upon further and further testing will be carried out to ensure this input system wont be broken.