TextToNumber Class Modifications
Implemented a method that will convert inputs that end in years,  weeks and days into months for the purposes of simulation input. Will adjust the class to ensure it returns the value as an integer. Added a display prompt to request a true or false value for if the user wants a monthly simulation report or not.